ticket VS-1120 add support for analyze and store ip encapsulation new confifuration parameters:  - save_ip_from_encaps_ipheader: yes/NO  - save_ip_from_encaps_ipheader_only_gre: yes/NO

+++ sniff_inline.h
@@ -29,7 +29,7 @@ inline
 #endif
 bool parseEtherHeader(int pcapLinklayerHeaderType, u_char* packet,
 		      sll_header *&header_sll, ether_header *&header_eth, u_char **header_ppp_o_e,
-		      u_int &header_ip_offset, int &protocol, u_int16_t &vlan);
+		      u_int16_t &header_ip_offset, u_int16_t &protocol, u_int16_t &vlan);
 
 #if SNIFFER_INLINE_FUNCTIONS
 inline
